Tester for å lese siste timestamp
